pyproject.toml
[project]
name = "ml-test"
version = "0.1.0"
description = "Test for ML job"
authors = [
    {name = "Mateus Furquim", email = "mfurquimdev@gmail.com"},
]
dependencies = [
    "pandas==1.5.3",
    "numpy==1.25.0",
    "scipy==1.11.3",
    "scikit-learn==1.3.1",
    "fastapi==0.104.1",
    "python-multipart==0.0.6",
    "httpx==0.25.1",
    "uvicorn>=0.24.0.post1",
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = {text = "MIT"}

